django-project/djangoRestFramework/REsFramework/requirements.txt
Django==1.11
django-filter==2.2.0
djangorestframework==3.11.0
Markdown==3.2.1
Pillow==7.1.1
pkg-resources==0.0.0
pytz==2019.3
